Introduction to Pseudocodes, Algorithms & Github
Algorithm is a process or set of rules to be followed in calculations or other problem-solving operations, especially by a computer. Pseudocode is a way of writing or describing programming code or algorithms in a natural language such as English. GitHub is a developer platform that allows developers to create, store, manage and share their code